chore: move "ws" to core root
